Maxbuf is unsigned long, suggested by Alexey Melnikov <Alexey.Melnikov@isode.com>.

+++ src/callbacks.h
@@ -57,7 +57,8 @@ client_callback_passcode (Gsasl_session_ctx * ctx,
 extern Gsasl_qop
 client_callback_qop (Gsasl_session_ctx * ctx, Gsasl_qop serverqops);
 
-extern int client_callback_maxbuf (Gsasl_session_ctx * ctx, int servermaxbuf);
+extern unsigned long client_callback_maxbuf (Gsasl_session_ctx * ctx,
+					     unsigned long servermaxbuf);
 
 extern int
 client_callback_realm (Gsasl_session_ctx * ctx, char *out, size_t * outlen);
@@ -73,7 +74,7 @@ server_callback_anonymous (Gsasl_session_ctx * ctx, const char *message);
 
 extern Gsasl_qop server_callback_qop (Gsasl_session_ctx * ctx);
 
-extern int server_callback_maxbuf (Gsasl_session_ctx * ctx);
+extern unsigned long server_callback_maxbuf (Gsasl_session_ctx * ctx);
 
 extern int
 server_callback_realm (Gsasl_session_ctx * ctx,
